refactor: Rework tabs / component layout
