+++ src/pages/screening/pupil/PupilDetail.tsx
@@ -1,45 +1,22 @@
 import { Button } from '@/components/Button';
-import { Checkbox } from '@/components/Checkbox';
-import { Label } from '@/components/Label';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/Panels';
 import { Typography } from '@/components/Typography';
-import { ExternalSchoolSearch, Gender_Enum as Gender, Pupil_Screening_Status_Enum } from '@/gql/graphql';
-import { asTranslationKey } from '@/helper/string-helper';
+import { Pupil_Screening_Status_Enum } from '@/gql/graphql';
 import { PupilForScreening, PupilScreening } from '@/types';
-import { getGradeLabel } from '@/Utility';
-import { EditGradeModal } from '@/widgets/screening/EditGradeModal';
-import { EditLanguagesModal } from '@/widgets/screening/EditLanguagesModal';
-import { EditSubjectsModal } from '@/widgets/screening/EditSubjectsModal';
-import { CheckedState } from '@radix-ui/react-checkbox';
-import { IconDeviceFloppy } from '@tabler/icons-react';
-import { useMemo, useState, useEffect } from 'react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ButtonField } from '../components/ButtonField';
-import { EditLocationModal } from '../components/EditLocationModal';
-import { EditSchoolTypeModal } from '../components/EditSchoolTypeModal';
-import { PupilHistory } from './PupilHistory';
-import { SchoolSearchInput } from '../components/SchoolSearchInput';
+import { PupilMatchingHistory, PupilScreeningsHistory } from './PupilHistory';
 import { ScreeningSuggestionCard } from '@/widgets/screening/ScreeningSuggestionCard';
 import { ScreenPupil } from './ScreenPupil';
 import { gql } from '@/gql';
 import { useMutation } from '@apollo/client';
 import { toast } from 'sonner';
-import { useRoles } from '@/hooks/useApollo';
+import PersonalDetails from './PersonalDetails';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/Panels';
 
 interface PupilDetailProps {
     pupil: PupilForScreening;
     refresh: () => Promise<void>;
 }
-
-const UPDATE_PUPIL_MUTATION = gql(`
-    mutation ScreenerUpdatePupil($pupilId: Float!, $data: PupilUpdateInput!) {
-        pupilUpdate(pupilId: $pupilId, data: $data)
-    }
-`);
-
-const CREATE_LOGIN_TOKEN_MUTATION = gql(`
-    mutation AdminAccess($userId: String!) { tokenCreateAdmin(userId: $userId) }
-`);
 
 const REQUEST_MATCH_MUTATION = gql(`
     mutation PupilRequestMatch($pupilId: Float!) { pupilCreateMatchRequest(pupilId: $pupilId) }
@@ -49,49 +26,11 @@ const REVOKE_MATCH_REQUEST_MUTATION = gql(`
     mutation PupilRevokeMatchRequest($pupilId: Float!) { pupilDeleteMatchRequest(pupilId: $pupilId) }
 `);
 
-interface FormErrors {
-    languages?: string;
-    grade?: string;
-    subjects?: string;
-}
-
 const PupilDetail = ({ pupil, refresh }: PupilDetailProps) => {
     const { t } = useTranslation();
-    const myRoles = useRoles();
-    const [showEditLocation, setShowEditLocation] = useState(false);
-    const [showEditSchoolType, setShowEditSchoolType] = useState(false);
-    const [showEditGrade, setShowEditGrade] = useState(false);
-    const [showEditSubjects, setShowEditSubjects] = useState(false);
-    const [showEditLanguages, setShowEditLanguages] = useState(false);
 
-    const [onlyMatchWithWomen, setOnlyMatchWithWomen] = useState<CheckedState>(pupil.onlyMatchWith === Gender.Female);
-    const [hasSpecialNeeds, setHasSpecialNeeds] = useState<CheckedState>(pupil.hasSpecialNeeds);
-    const [pupilLocation, setPupilLocation] = useState(pupil.state);
-    const [schoolType, setSchoolType] = useState(pupil.schooltype);
-    const [grade, setGrade] = useState(pupil.gradeAsInt);
-    const [subjects, setSubjects] = useState(pupil.subjectsFormatted);
-    const [languages, setLanguages] = useState(pupil.languages);
-    const [school, setSchool] = useState<ExternalSchoolSearch | undefined>(pupil.school as any);
-    const [mutationUpdatePupil, { loading: isUpdating }] = useMutation(UPDATE_PUPIL_MUTATION);
-    const [mutationCreateLoginToken] = useMutation(CREATE_LOGIN_TOKEN_MUTATION);
     const [mutationRequestMatch, { loading: isRequestingMatch }] = useMutation(REQUEST_MATCH_MUTATION);
     const [mutationRevokeMatchRequest, { loading: isRevokingMatchRequest }] = useMutation(REVOKE_MATCH_REQUEST_MUTATION);
-
-    const [errors, setErrors] = useState<FormErrors>({});
-
-    useEffect(() => {
-        const updatedErrors = {};
-        if (!languages || languages.length === 0) {
-            setErrors({ ...updatedErrors, languages: t('screening.errors.language_missing') });
-        }
-        if (grade === null || grade === undefined) {
-            setErrors({ ...updatedErrors, grade: t('screening.errors.grade_missing') });
-        }
-        if (!subjects || subjects.length === 0) {
-            setErrors({ ...updatedErrors, subjects: t('screening.errors.subjects_missing') });
-        }
-        setErrors(updatedErrors);
-    }, [languages, grade, subjects, t]);
 
     const { previousScreenings, screeningToEdit } = useMemo(() => {
         const previousScreenings: PupilScreening[] = [...pupil!.screenings!];
@@ -116,48 +55,6 @@ const PupilDetail = ({ pupil, refresh }: PupilDetailProps) => {
         // or in case the previous screening was already invalidated
         previousScreenings[0].invalidated;
 
-    const handleOnSavePupil = async () => {
-        try {
-            await mutationUpdatePupil({
-                variables: {
-                    pupilId: pupil.id,
-                    data: {
-                        gradeAsInt: grade,
-                        subjects: subjects.map((e) => ({ name: e.name, grade: e.grade, mandatory: e.mandatory })),
-                        languages: languages as any,
-                        onlyMatchWith: onlyMatchWithWomen === true ? Gender.Female : (undefined as any),
-                        hasSpecialNeeds: hasSpecialNeeds === true,
-                        school: {
-                            name: school?.name,
-                            schooltype: schoolType as any,
-                            state: pupilLocation as any,
-                        },
-                    },
-                },
-            });
-            toast.success(t('changesWereSaved'));
-        } catch (error) {
-            toast.success(t('error'));
-        }
-    };
-
-    const impersonate = async () => {
-        // We need to work around the popup blocker of modern browsers, as you can only
-        // call window.open(.., '_blank') in a synchronous event handler of onClick,
-        // so we open the window before we call any asynchronous functions and later set the URL when we have the data.
-        const w = window.open('', '_blank');
-        if (w != null) {
-            const res = await mutationCreateLoginToken({ variables: { userId: `pupil/${pupil!.id}` } });
-            const token = res?.data?.tokenCreateAdmin;
-
-            w.location.href =
-                process.env.NODE_ENV === 'production'
-                    ? `https://app.lern-fair.de/login-token?secret_token=${token}&temporary`
-                    : `http://localhost:3000/login-token?secret_token=${token}&temporary`;
-            w.focus();
-        }
-    };
-
     const handleOnRequestMatch = async () => {
         try {
             await mutationRequestMatch({ variables: { pupilId: pupil.id } });
@@ -178,150 +75,88 @@ const PupilDetail = ({ pupil, refresh }: PupilDetailProps) => {
         refresh();
     };
 
-    const handleOnSelectSchool = (school: ExternalSchoolSearch) => {
-        setSchool(school);
-        if (school.schooltype) {
-            setSchoolType(school.schooltype as any);
-        }
-        if (school.state) {
-            setPupilLocation(school.state as any);
-        }
-    };
-
     return (
         <div className="mt-8">
-            <Tabs defaultValue="personal">
-                <TabsList className="max-h-9 p-1">
-                    <TabsTrigger className="max-h-7" value="personal">
-                        Profil
+            <Tabs defaultValue="main">
+                <TabsList className="max-h-9 p-1 mb-2">
+                    <TabsTrigger className="max-h-7" value="main">
+                        Profil + Screening
                     </TabsTrigger>
-                    <TabsTrigger className="max-h-7" value="screening">
-                        Screening
+                    <TabsTrigger className="max-h-7" value="matching">
+                        Matching
                     </TabsTrigger>
-                    <TabsTrigger className="max-h-7" value="history">
-                        Verlauf
+                    <TabsTrigger className="max-h-7" value="recommendation">
+                        Empfehlungen
                     </TabsTrigger>
                 </TabsList>
-                <TabsContent value="personal">
-                    <div className="mt-5">
+                <TabsContent value="main">
+                    <div className="shadow-md px-6 py-8 rounded-md">
                         <Typography variant="h4" className="mb-5">
                             Persönliche Daten
                         </Typography>
-                        <div className="flex flex-wrap gap-6">
-                            <SchoolSearchInput onSelect={handleOnSelectSchool} defaultValue={school?.name} />
-                            <div className="flex flex-col gap-y-2">
-                                <ButtonField label="Schulort" onClick={() => setShowEditLocation(true)}>
-                                    {t(`lernfair.states.${pupilLocation}`) ?? 'Schulort bearbeiten'}
-                                </ButtonField>
-                            </div>
-                            <div className="flex flex-col gap-y-2">
-                                <ButtonField label="Schulform" onClick={() => setShowEditSchoolType(true)}>
-                                    {t(`lernfair.schooltypes.${schoolType}`) ?? 'Schulform bearbeiten'}
-                                </ButtonField>
-                            </div>
-                            <div className="flex flex-col gap-y-2">
-                                <ButtonField label="Klassenstufe" onClick={() => setShowEditGrade(true)}>
-                                    {getGradeLabel(grade) ?? 'Klassenstufe bearbeiten'}
-                                </ButtonField>
-                                <Typography variant="sm" className="text-destructive">
-                                    {errors.grade}
-                                </Typography>
-                            </div>
-                            <div className="flex flex-col gap-y-2">
-                                <ButtonField className="min-w-[350px]" label="Fächer" onClick={() => setShowEditSubjects(true)}>
-                                    {subjects.map((e) => t(asTranslationKey(`lernfair.subjects.${e.name}`))).join(', ') ?? 'Fächer bearbeiten'}
-                                </ButtonField>
-                                <Typography variant="sm" className="text-destructive">
-                                    {errors.subjects}
-                                </Typography>
-                            </div>
-                            <div className="flex flex-col gap-y-2">
-                                <ButtonField className="min-w-[350px]" label="Gesprochene Sprachen" onClick={() => setShowEditLanguages(true)}>
-                                    {languages.map((e) => t(asTranslationKey(`lernfair.languages.${e.toLowerCase()}`))).join(', ') ?? 'Sprachen bearbeiten'}
-                                </ButtonField>
-                                <Typography variant="sm" className="text-destructive">
-                                    {errors.languages}
-                                </Typography>
-                            </div>
-                            <div className="flex gap-x-7 mt-2">
-                                <div className="flex gap-x-2 items-center">
-                                    <Checkbox id="onlyMatchWith" checked={onlyMatchWithWomen} onCheckedChange={setOnlyMatchWithWomen} />{' '}
-                                    <Label htmlFor="onlyMatchWith">Nur mit Frauen matchen</Label>
-                                </div>
-                                <div className="flex gap-x-2 items-center">
-                                    <Checkbox id="specialNeeds" checked={hasSpecialNeeds} onCheckedChange={setHasSpecialNeeds} />{' '}
-                                    <Label htmlFor="specialNeeds">Besonderer Anspruch</Label>
-                                </div>
-                            </div>
-                        </div>
-                        <div className="mt-10 flex items-center gap-x-4">
-                            <Button onClick={handleOnSavePupil} isLoading={isUpdating} leftIcon={<IconDeviceFloppy />} className="w-80">
-                                Speichern
-                            </Button>
-                            {myRoles.includes('TRUSTED_SCREENER') && pupil.active && (
-                                <Button variant="outline" onClick={impersonate}>
-                                    Als Nutzer anmelden
-                                </Button>
-                            )}
-                        </div>
-                        <div className="mt-10 flex flex-col">
-                            <Typography variant="h4" className="mb-5">
-                                Matching
-                            </Typography>
-                            <div className="flex flex-col gap-y-4">
-                                <Typography className="block">{pupil.openMatchRequestCount} Matchanfragen</Typography>
-                                <div className="flex items-center gap-x-4">
-                                    <Button
-                                        isLoading={isRequestingMatch}
-                                        disabled={needsScreening && !screeningToEdit}
-                                        reasonDisabled="Zuerst muss ein Screening angelegt werden"
-                                        onClick={handleOnRequestMatch}
-                                    >
-                                        Match anfragen
-                                    </Button>
-                                    <Button
-                                        variant="outline"
-                                        isLoading={isRevokingMatchRequest}
-                                        disabled={pupil.openMatchRequestCount === 0}
-                                        reasonDisabled="Keine offene Matchanfrage"
-                                        onClick={handleOnRemoveMatchRequest}
-                                    >
-                                        Anfrage zurücknehmen
-                                    </Button>
-                                </div>
-                            </div>
-                        </div>
-                        <div className="mt-10 flex flex-col">
-                            <Typography variant="h4" className="mb-5">
-                                Empfehlungen
-                            </Typography>
-                            <ScreeningSuggestionCard userID={`pupil/${pupil.id}`} />
-                        </div>
+                        <PersonalDetails pupil={pupil} refresh={refresh} />
                     </div>
-                </TabsContent>
-                <TabsContent value="screening">
-                    <div className="mt-5">
+                    <div className="shadow-md px-6 py-8 rounded-md mt-10">
                         <Typography variant="h4" className="mb-5">
                             Screening
                         </Typography>
                         <ScreenPupil pupil={pupil} screening={screeningToEdit ?? undefined} needsScreening={needsScreening} refresh={refresh} />
                     </div>
+                    {previousScreenings.length > 0 && (
+                        <div className="shadow-md px-6 py-8 rounded-md mt-10">
+                            <Typography variant="h4" className="mb-5">
+                                Vorherige Screenings
+                            </Typography>
+                            <PupilScreeningsHistory screenings={previousScreenings} />
+                        </div>
+                    )}
                 </TabsContent>
-                <TabsContent value="history">
-                    <div className="mt-5">
+                <TabsContent value="matching">
+                    <div className="flex flex-col shadow-md px-6 py-8 rounded-md">
                         <Typography variant="h4" className="mb-5">
-                            Verlauf
+                            Match Anfragen
                         </Typography>
-                        <PupilHistory pupil={pupil} previousScreenings={previousScreenings} />
+                        <div className="flex flex-col gap-y-4">
+                            <Typography className="block">{pupil.openMatchRequestCount} Matchanfragen</Typography>
+                            <div className="flex items-center gap-x-4">
+                                <Button
+                                    isLoading={isRequestingMatch}
+                                    disabled={needsScreening && !screeningToEdit}
+                                    reasonDisabled="Zuerst muss ein Screening angelegt werden"
+                                    onClick={handleOnRequestMatch}
+                                >
+                                    Match anfragen
+                                </Button>
+                                <Button
+                                    variant="outline"
+                                    isLoading={isRevokingMatchRequest}
+                                    disabled={pupil.openMatchRequestCount === 0}
+                                    reasonDisabled="Keine offene Matchanfrage"
+                                    onClick={handleOnRemoveMatchRequest}
+                                >
+                                    Anfrage zurücknehmen
+                                </Button>
+                            </div>
+                        </div>
+                    </div>
+                    {pupil.matches && pupil.matches.length > 0 && (
+                        <div className="flex flex-col shadow-md px-6 py-8 rounded-md mt-10">
+                            <Typography variant="h4" className="mb-5">
+                                Matching Verlauf
+                            </Typography>
+                            <PupilMatchingHistory matches={pupil.matches} />
+                        </div>
+                    )}
+                </TabsContent>
+                <TabsContent value="recommendation">
+                    <div className="flex flex-col shadow-md px-6 py-8 rounded-md">
+                        <Typography variant="h4" className="mb-5">
+                            Empfehlungen
+                        </Typography>
+                        <ScreeningSuggestionCard userID={`pupil/${pupil.id}`} />
                     </div>
                 </TabsContent>
             </Tabs>
-
-            <EditLocationModal state={pupilLocation} onSave={setPupilLocation} isOpen={showEditLocation} onOpenChange={setShowEditLocation} />
-            <EditSchoolTypeModal schoolType={schoolType} onSave={setSchoolType} isOpen={showEditSchoolType} onOpenChange={setShowEditSchoolType} />
-            <EditGradeModal grade={grade} onSave={setGrade} onOpenChange={setShowEditGrade} isOpen={showEditGrade} />
-            <EditSubjectsModal type="pupil" subjects={subjects} onSave={setSubjects} onOpenChange={setShowEditSubjects} isOpen={showEditSubjects} />
-            <EditLanguagesModal languages={languages} onSave={setLanguages} onOpenChange={setShowEditLanguages} isOpen={showEditLanguages} />
         </div>
     );
 };
